# doc/api.yaml
openapi: 3.0.0
info:
  title: Trello Service
  description: Let's try to create a competitor for Trello!
  version: 1.0.0
servers:
  - url: /
components:
  schemas:
...
    Task:
      required:
        - title
      properties:
        id:
          type: string
        title:
          type: string
        order:
          type: number
        description:
          type: string
        userId:
          type: string
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /boards/{boardId}/tasks:
    parameters:
      - name: boardId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Tasks
      summary: Get Tasks by boardId
      description: Gets tasks by the Board ID
        (e.g. “/board/1/tasks”)
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        401:
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Tasks
      summary: Create new task
      description: Creates a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                title:
                  type: string
                  description: The task's title
                order:
                  type: integer
                  description: The task's order
                description:
                  type: string
                userId:
                  type: string
                  description: The task's Assignee
                boardId:
                  type: string
                  description: Board ID
                columnId:
                  type: string
                  description: Column ID
              required:
                - title
                - order
                - boardId
                - columnId
      responses:
        201:
          description: The task has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /boards/{boardId}/tasks/{taskId}:
    parameters:
      - name: boardId
        in: path
        required: true
        schema:
          type: string
      - name: taskId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Tasks
      summary: Get Task by boardId and taskId
      description: Gets the Task by the Board's and task ID
        (e.g. “/board/1/tasks/123”)
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        401:
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Tasks
      summary: Update Task
      description: Updates the Task by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                title:
                  type: string
                  description: The task's title
                order:
                  type: integer
                  description: The task's order
                description:
                  type: string
                userId:
                  type: string
                  description: The task's Assignee
                boardId:
                  type: string
                  description: Board ID
                columnId:
                  type: string
                  description: Column ID
              required:
                - order
                - boardId
                - columnId
      responses:
        200:
          description: The task has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Task not found
    delete:
      tags:
        - Tasks
      summary: Delete Task
      description: Deletes Task by ID.
      responses:
        204:
          description: The task has been deleted
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Task not found
...